projects/target_endpoint_pythonsdkapp_addedfourthAPI/requirements.txt
requests==2.18.4
flask==2.0.1
faker==8.11.0
python-dotenv~=0.19.0